В первом приближении набросал CRUD для Organization через web, добавил линки для запроса одной организации. TODO: Добавить остальные линки, response.

+++ src/main/java/pro/b2b2b/germ/repository/OrganizationRepository.java
@@ -3,5 +3,5 @@ package pro.b2b2b.germ.repository;
 import org.springframework.data.repository.CrudRepository;
 import pro.b2b2b.germ.model.Organization;
 
-public interface OrganizationRepository extends CrudRepository<Organization, Integer> {
+public interface OrganizationRepository extends CrudRepository<Organization, Long> {
 }
